Fix url schema. Add /bookmarks/{id} endpoint to get a specific bookmark. Changed ddlAuto strategy to update rather than recreate DB every time.

+++ src/main/java/bookmarkapp/repositories/BookmarkRepository.java
@@ -3,5 +3,5 @@ package bookmarkapp.repositories;
 import bookmarkapp.models.Bookmark;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookmarkRepository extends CrudRepository<Bookmark, Long> {
+public interface BookmarkRepository extends CrudRepository<Bookmark, Integer> {
 }
